feat: adding the validation to the ownership of the account before any withrawal or deposit

+++ src/models/SavingAccount.java
@@ -1,11 +1,11 @@
 package models;
 import abstracts.Account;
-
+import java.math.BigDecimal;
 public class SavingAccount extends Account {
 
     private double interestRate;
 
-    public SavingAccount(String accountCode, int ownerId, double balance, double interestRate) {
+    public SavingAccount(String accountCode, int ownerId, BigDecimal balance, double interestRate) {
         super(accountCode, ownerId, balance); // call parent constructor
         this.interestRate = interestRate;
     }
